requirements.txt
Flask>=2.2,<2.3
Jinja2>=3.0,<3.1
werkzeug>=2.2,<2.3
pytest==6.2.5
psycopg2-binary
gunicorn==23.0.0